+++ mediation/LiftoffMonetize/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Liftoff Monetize Unity Mediation Plugin Changelog
 
-#### Next Version
+#### [Version 5.1.0](https://dl.google.com/googleadmobadssdk/mediation/unity/liftoffmonetize/LiftoffMonetizeUnityAdapter-5.1.0.zip)
+- Fixed a `NullPointerException` that caused a build failure when exporting as an Android project.
 - Updated the privacy APIs:
   * Removed the `UpdateConsentStatus()` and `UpdateCCPAStatus()` methods.
   * Removed the `VungleConsentStatus` and `VungleCCPAStatus` enums.
@@ -8,6 +9,9 @@
 - Removed the following methods from the `LiftoffMonetizeMediationExtras` class:
   * `SetAllPlacements(String[])`
   * `SetSoundEnabled(bool)`
+- Supports [Liftoff Monetize Android adapter version 7.3.1.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/liftoffmonetize/CHANGELOG.md#version-7310).
+- Supports [Liftoff Monetize iOS adapter version 7.3.1.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/LiftoffMonetize/CHANGELOG.md#version-7310).
+- Built and tested with the Google Mobile Ads Unity Plugin version 9.1.0.
 
 #### [Version 5.0.0](https://dl.google.com/googleadmobadssdk/mediation/unity/liftoffmonetize/LiftoffMonetizeUnityAdapter-5.0.0.zip)
 - Fixed a duplicate definition warning for `MediationExtras`.